fix errors

+++ 0x13-more_singly_linked_lists/8-sum_listint.c
@@ -9,7 +9,7 @@
 int sum_listint(listint_t *head)
 {
 	listint_t *current = head;
-	size_t sum = 0;
+	int sum = 0;
 
 	if (!head)
 		return (0);
